backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
psycopg2==2.9.6
bitcoinaddress==0.1.5
mnemonic==0.20
python-dotenv==1.0.0
tronpy==0.4.0
web3==6.1.0
